requirements.txt
asgiref==3.7.2
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.12.4
inflection==0.5.1
packaging==23.2
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.5
